Fixed functions of libft after correction by Moulinette. Added functions created for the ft_printf project. Delete ft_printf() because cub3d doesn't allow va_arg functions

+++ src/conversion/ft_atoi.c
@@ -3,14 +3,14 @@
 /*                                                        :::      ::::::::   */
 /*   ft_atoi.c                                          :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: dda-silv <dda-silv@student.42.fr>          +#+  +:+       +#+        */
+/*   By: dda-silv <dda-silv@student.42lisboa.com>   +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/01/11 18:44:56 by dda-silv          #+#    #+#             */
-/*   Updated: 2021/01/12 18:14:19 by dda-silv         ###   ########.fr       */
+/*   Updated: 2021/02/23 17:58:02 by dda-silv         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int		ft_atoi(const char *str)
+long long	ft_atoi(const char *str)
 {
 	int number;
 	int sign;
